requirements.txt
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
transformers>=4.35.0
torch>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0 